+++ android/libui/src/main/java/com/energy/view/tempcanvas/PointDraw.kt
@@ -21,20 +21,20 @@ class PointDraw(context: Context) : BaseDraw(context) {
         private const val TOUCH_EXTRA = 20f
     }
 
-    private val textPointMargin = ScreenUtils.dp2px(4).toFloat()
-    private val labelPointMargin = ScreenUtils.dp2px(24).toFloat()
+    private val textPointMargin = ScreenUtils.dp2px(4f).toFloat()
+    private val labelPointMargin = ScreenUtils.dp2px(24f).toFloat()
     private val pointList = LinkedList<PointView>()
     private var tempPoint: PointView? = null
     
     private val textPaint = Paint().apply {
-        strokeWidth = ScreenUtils.dp2px(STROKE_WIDTH).toFloat()
-        textSize = ScreenUtils.sp2px(TEXT_SIZE).toFloat()
+        strokeWidth = ScreenUtils.dp2px(STROKE_WIDTH.toFloat()).toFloat()
+        textSize = ScreenUtils.sp2px(TEXT_SIZE.toFloat()).toFloat()
         color = Color.WHITE
         textAlign = Paint.Align.CENTER
     }
     
     private val bgPaint = TextPaint().apply {
-        strokeWidth = ScreenUtils.dp2px(1).toFloat()
+        strokeWidth = ScreenUtils.dp2px(1f).toFloat()
     }
     
     private var fontMetrics: Paint.FontMetrics? = null
